Made it work
Import of union from set to get a way to get unique location on the
image for the find-colour-block-in (this can probably be done a better
way)

discovered that quote in front of [] makes a symbol vector.

piet-push now converts from lazy-list

piet-update now takes how many elements it should take from value and
machine is first argument; updated for all functions as well.

rotate did it the wrong way for positive numbers I think. Now it does
it correctly. This function could be revised.

roll now makes sure it does not leave a lazy-list

now there is big out text that shows it has been printed out in the outs

added piet-nop ; no operation

grab-pixels converts the image into a format the program understands

find-colour-block-in finds the full colour block for the current
colour; much room for improvement here.

choose-codel chooses which codel to go to next.

piet-commands finds the command that will be run on the machine.

piet-interpreter interprets the image given the size of each codel.