Fixed a lot of bad code
piet-machine now also has an :out, it is where the machine writes to
out.

:value is now a single number

Removed piet from function names.

push no longer has a need to remove anything in :value. It only pushes
the :value onto the stack.

pop is no longer used to move to value, only removing the top element
if it is called from the interpreter.

Using `nthnext` and `take n` inside piet-update to remove and calculate
on the stack elements.

need to use clojure.core/mod to call mod as mod now is piet-mod instead.

roll has a very large comp-partial-juxt function that could probably be
shorter, but I had fun making it :D

the different out-functions now adds to the :out key in machine.

Using #' in commands to be able to access the metadata later.

Moved to assoc where it wasn't neccessary with update-in

Now prints the output inside the machine at the end.

Need read-string on codel-size since it's from input and it will be a
string.